# api-docs.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Register an account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEnvelopeDto"
                }
              }
            }
          }
        },
        "summary": "Log in",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Log out",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "getSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "hungr_session": []
          }
        ],
        "summary": "Get the authenticated session",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "hungr_session": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "updateCurrentUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "hungr_session": []
          }
        ],
        "summary": "Update current user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/foods": {
      "get": {
        "operationId": "listFoods",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodListEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "hungr_session": []
          }
        ],
        "summary": "List your foods",
        "tags": [
          "Foods"
        ]
      },
      "post": {
        "operationId": "createFood",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFoodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "hungr_session": []
          }
        ],
        "summary": "Create a food",
        "tags": [
          "Foods"
        ]
      }
    },
    "/api/v1/foods/{id}": {
      "get": {
        "operationId": "getFood",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "hungr_session": []
          }
        ],
        "summary": "Get one of your foods",
        "tags": [
          "Foods"
        ]
      },
      "patch": {
        "operationId": "updateFood",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFoodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "hungr_session": []
          }
        ],
        "summary": "Update your food",
        "tags": [
          "Foods"
        ]
      },
      "delete": {
        "operationId": "deleteFood",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "hungr_session": []
          }
        ],
        "summary": "Delete your food",
        "tags": [
          "Foods"
        ]
      }
    },
    "/healthz": {
      "get": {
        "operationId": "getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "API and MongoDB are healthy.\n\nThe Health Check is successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {},
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "MongoDB is unavailable.\n\nThe Health Check is not successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        },
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Check API health",
        "tags": [
          "Status"
        ]
      }
    }
  },
  "info": {
    "title": "Hungr API",
    "description": "Private food sharing API for the Hungr application.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hungr_session"
      }
    },
    "schemas": {
      "CredentialsDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "you@example.com"
          },
          "password": {
            "type": "string",
            "example": "correct-horse-battery-staple",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "email": {
            "type": "string",
            "example": "you@example.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "UserEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UserDataDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "new-email@example.com"
          }
        }
      },
      "FoodResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "name": {
            "type": "string",
            "example": "Tacos"
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/tacos.jpg",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "imageUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "FoodListDataDto": {
        "type": "object",
        "properties": {
          "foods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FoodResponseDto"
            }
          }
        },
        "required": [
          "foods"
        ]
      },
      "FoodListEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FoodListDataDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateFoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tacos",
            "maxLength": 120
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/tacos.jpg",
            "nullable": true
          }
        },
        "required": [
          "name"
        ]
      },
      "FoodDataDto": {
        "type": "object",
        "properties": {
          "food": {
            "$ref": "#/components/schemas/FoodResponseDto"
          }
        },
        "required": [
          "food"
        ]
      },
      "FoodEnvelopeDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FoodDataDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateFoodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tacos",
            "maxLength": 120
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/tacos.jpg",
            "nullable": true
          }
        }
      }
    }
  }
}