# apps/api/src/auth/dto/credentials.dto.ts
import { ApiProperty } from "@nestjs/swagger";
import { IsEmail, IsString, MinLength } from "class-validator";

export class CredentialsDto {
  @ApiProperty({ example: "you@example.com" })
  @IsEmail()
  email: string;

  @ApiProperty({ example: "correct-horse-battery-staple", minLength: 8 })
  @IsString()
  @MinLength(8)
  password: string;
}